# linear-clone-backend-js/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Linear Clone API
  description: >
    Backend API for a Linear-inspired project management application.

    This API provides authentication, user profiles, organizations, teams,

    issues, comments, labels, watchers, workflows, projects, cycles, and
    notifications.


    ## Authentication

    Most endpoints require a Bearer JWT access token in the `Authorization`
    header.

    Obtain tokens via `POST /api/v1/auth/login` or `POST /api/v1/auth/register`.

    The refresh token is stored in an HttpOnly cookie (`refreshToken`) scoped to
    `/api/v1/auth/refresh` with `SameSite=Strict`. The POST /api/v1/auth/refresh
    endpoint reads the cookie first, falling back to a request body field for
    backward compatibility. The cookie is set on login/refresh and cleared on
    logout.


    ## WebSocket Gateway

    Real-time events are available at `ws://<host>/ws`. After connecting, send
    an

    `authenticate` message with your access token, then subscribe to channels

    (`team:<id>`, `issue:<id>`, `user:<id>`) to receive live updates.
  version: 1.0.0
  contact:
    email: dev@example.com
  license:
    name: ISC
...
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      deprecated: false
      description: >-
        Creates a new user account and returns JWT tokens. The refresh token is
        set as an HttpOnly cookie (`refreshToken`) scoped to `/api/v1/auth/refresh`
        with `SameSite=Strict`. The refresh token is not returned in the response
        body.
      operationId: authRegister
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
                - password
              properties:
                email:
                  type: string
                  format: email
                  examples:
                    - user@example.com
                name:
                  type: string
                  minLength: 1
                  examples:
                    - John Doe
                password:
                  type: string
                  format: password
                  minLength: 8
                  examples:
                    - securePassword123
            example:
              email: jlagoria@example.com
              name: John Lagoria
              password: securePassword#123
        required: true
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              description: >-
                `refreshToken=<token>; HttpOnly; Secure; SameSite=Strict;
                Path=/api/v1/auth/refresh; Max-Age=604800`. The refresh token is
                not returned in the response body.
              schema:
                type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          headers: {}
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
          headers: {}
      security: []
  /api/v1/auth/login:
    post:
      summary: Authenticate user and get tokens
      deprecated: false
      description: >-
        Authenticates a user and returns JWT tokens. The refresh token is set as
        an HttpOnly cookie (`refreshToken`) scoped to `/api/v1/auth/refresh` with
        `SameSite=Strict`. Supports "Remember Me" for extended refresh token
        lifetime (30-day cookie Max-Age).
      operationId: authLogin
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  examples:
                    - user@example.com
                  description: User email
                password:
                  type: string
                  format: password
                  examples:
                    - securePassword123
                  description: User password
                rememberMe:
                  type: boolean
                  default: false
                  description: Extends refresh token to 30 days
            example:
              email: jlagoria@example.com
              password: securePassword#123
              rememberMe: true
        required: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/UserObject'
                      accessToken:
                        type: string
                    required:
                      - user
                      - accessToken
          headers:
            Set-Cookie:
              description: >-
                `refreshToken=<token>; HttpOnly; Secure; SameSite=Strict;
                Path=/api/v1/auth/refresh; Max-Age=604800` (or 2592000 with
                `rememberMe`). The refresh token is not returned in the response
                body.
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ValidationError'
          description: Validation error
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
          headers: {}
      security: []
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token using a refresh token
      deprecated: false
      description: >-
        Exchanges a valid refresh token for a new access token. The refresh
        token is read from the `refreshToken` HttpOnly cookie (preferred) or
        from the request body (fallback for backward compatibility). The old
        refresh token is invalidated (single-use rotation) and a new one is set
        as a cookie.
      operationId: authRefreshToken
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: >-
                    Optional body fallback. Only used if the `refreshToken`
                    cookie is absent.
            example:
              refreshToken: '{{refreshToken}}'
        required: false
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                    required:
                      - accessToken
                required:
                  - data
              example:
                data:
                  accessToken: >-
                    eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxYjkxZGFkMC05ZDA0LTQyMjMtOGZlZi0yMGY0Y2ZhMDI1NzkiLCJ0eXBlIjoiYWNjZXNzIiwic2lkIjoiMDM5NmEzZWMtNWE4Yy00NzEwLTg5MGUtNDMxYjdkZTI4ZDY4IiwiaWF0IjoxNzg0MjQwNDg4LCJleHAiOjE3ODQyNDEzODgsImp0aSI6IjA5YTJmZTU5LTYxMGItNGI5Mi04Yzk4LWQzYzZlZjEwOWNhMSJ9.Db1HtgOB9NmaGqaJjrPJ5f3PyQTPJxdiV0mEni_oi4E
          headers:
            Set-Cookie:
              description: >-
                `refreshToken=<new-token>; HttpOnly; Secure; SameSite=Strict;
                Path=/api/v1/auth/refresh; Max-Age=604800`. The refresh token is
                rotated and set as a cookie; it is not returned in the response
                body.
              schema:
                type: string
        '401':
          description: Token expired or revoked
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenExpiredError'
                  - $ref: '#/components/schemas/TokenRevokedError'
          headers: {}
        '422':
          $ref: '#/components/responses/ValidationFailed'
          description: Validation failed
      security: []
  /api/v1/auth/logout:
    post:
      summary: Logout and revoke current session
      deprecated: false
      description: >-
        Terminates the user's session and invalidates all refresh tokens.
        Clears the `refreshToken` HttpOnly cookie. Requires a valid access
        token. Idempotent — calling multiple times returns success.
      operationId: authLogout
      tags:
        - Authentication
      parameters: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                        enum:
                          - true
          headers:
            Set-Cookie:
              description: >-
                `refreshToken=; HttpOnly; Secure; SameSite=Strict;
                Path=/api/v1/auth/refresh; Max-Age=0`. The cookie is cleared on
                logout.
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication required
      security:
        - BearerAuth: []
  /api/v1/auth/sessions:
    get:
      summary: List all active sessions for the current user
      deprecated: false
      description: ''
      operationId: listSessions
      tags:
        - Authentication
        - Sessions
      parameters: []
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      sessions:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            ipAddress:
                              type: string
                            userAgent:
                              type: string
                            rememberMe:
                              type: boolean
                            createdAt:
                              type: string
                            lastActivityAt:
                              type: string
                            isCurrent:
                              type: boolean
                    required:
                      - sessions
                required:
                  - data
              example:
                data:
                  sessions:
                    - id: 75d7053e-ce90-4b1a-9d4d-e24bb417d4b3
                      ipAddress: 127.0.0.1
                      userAgent: Apidog/1.0.0 (https://apidog.com)
                      rememberMe: true
                      createdAt: '2026-07-16T22:22:25.507Z'
                      lastActivityAt: '2026-07-16T22:22:25.507Z'
                      isCurrent: true
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication required
      security:
        - BearerAuth: []
  /api/v1/auth/sessions/revoke-all:
    post:
      summary: Revoke all sessions except the current one
      deprecated: false
      description: >-
        Revokes all sessions except the current one, effectively signing out
        from all other devices while remaining authenticated on the current
        device.
      operationId: revokeAllSessions
      tags:
        - Authentication
        - Sessions
      parameters: []
      responses:
        '200':
          description: Sessions revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                        enum:
                          - true
                      revokedCount:
                        type: integer
                        minimum: 0
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication required
      security:
        - BearerAuth: []
  /api/v1/auth/sessions/{sessionId}:
    delete:
      summary: Revoke a specific session
      deprecated: false
      description: >-
        Revokes (terminates) a specific session by its identifier. Revoking the
        current session is equivalent to logging out.
      operationId: revokeSession
      tags:
        - Authentication
        - Sessions
      parameters:
        - name: sessionId
          in: path
          description: ''
          required: true
          example: 75d7053e-ce90-4b1a-9d4d-e24bb417d4b3
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                        enum:
                          - true
          headers: {}
        '400':
          $ref: '#/components/responses/ValidationError'
          description: Validation error
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication required
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
          headers: {}
      security:
        - BearerAuth: []
...
components:
  schemas:
    AuthResponse:
      type: object
      required:
        - user
        - accessToken
      properties:
        user:
          $ref: '#/components/schemas/UserObject'
        accessToken:
          type: string
    UserObject:
      type: object
      required:
        - id
        - email
        - name
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    UnauthorizedError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - UNAUTHORIZED
            message:
              type: string
...
    NotFoundError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - NOT_FOUND
            message:
              type: string
    ConflictError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - CONFLICT
            message:
              type: string
    TokenExpiredError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - TOKEN_EXPIRED
            message:
              type: string
    TokenRevokedError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - TOKEN_REVOKED
            message:
              type: string
...
    ValidationFailed:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_FAILED
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationFailed'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or register
servers: []
security: []