# linear-clone-backend-js/src/modules/auth/adapters/in/dto.ts
import { z } from 'zod';

// Request schemas
export const RegisterRequestSchema = z.object({
  email: z.string().email('Invalid email format'),
  name: z.string().min(1, 'Name is required'),
  password: z.string().min(8, 'Password must be at least 8 characters'),
});

export const LoginRequestSchema = z.object({
  email: z.string().email('Invalid email format'),
  password: z.string().min(1, 'Password is required'),
  rememberMe: z.boolean().optional().default(false),
});

// Response schemas
export const UserObjectSchema = z.object({
  id: z.string().uuid(),
  email: z.string().email(),
  name: z.string(),
  createdAt: z.date().optional(),
});

export const AuthResponseSchema = z.object({
  data: z.object({
    user: UserObjectSchema,
    accessToken: z.string(),
  }),
});

export const RefreshRequestSchema = z.object({
  refreshToken: z.string().optional(),
});

export const RefreshResponseSchema = z.object({
  data: z.object({
    accessToken: z.string(),
  }),
});

export const ErrorResponseSchema = z.object({
  error: z.object({
    code: z.string(),
    message: z.string(),
    details: z
      .array(
        z.object({
          field: z.string(),
          message: z.string(),
        }),
      )
      .optional(),
  }),
});

// Types
export type RegisterRequest = z.infer<typeof RegisterRequestSchema>;
export type LoginRequest = z.infer<typeof LoginRequestSchema>;
export type UserObject = z.infer<typeof UserObjectSchema>;
export type AuthResponse = z.infer<typeof AuthResponseSchema>;
export type RefreshRequest = z.infer<typeof RefreshRequestSchema>;
export type RefreshResponse = z.infer<typeof RefreshResponseSchema>;
export type ErrorResponse = z.infer<typeof ErrorResponseSchema>;